refactor(ai): consolidate Responses implementation
